requirements.txt
Flask
openai==1.97.1
python-dotenv
